requirements.txt
Flask
huggingface_hub
llama-cpp-python